adapter: Refactor device in separate class
